Refactor function placement
